Enrich all tracking events with view specific meta (#4412)
# Why

For the `page.view` event we introduced some default tracking meta
data properties which are considered generally useful. There is now
desire to have the same meta data fields on all Mixpanel events.

# What

Apply the same meta data to all Mixpanel events. To do so we have
merge the `tracking.js` and `viewTracking.js` files. This makes it
easier to keep track of the global tracking state in one place.

To keep `tracking.js` manageable in size the tracking event names
were moved to a separate file named `eventNames.js`.