QueryBuilder tracking code, mixpanel events for infra explore (#4494)
* refactor: more descriptive function name for url setter

* adds filter added, removed ,cleared event notifications from QueryBuilder and uses them in infra explore

* emit groupAdded and groupRemoved events from the grouping configurator component and use them in infra explore mixpanel tracking

* adds infra.explore.navigate.to.entity event

* adds tracking event for type selector in infra explore

* tracks infra.explore.group.focused.on events

* track infra.explore.group.{collapsed, expanded} events for mix panel

* first pass at infra.explore.load.more event with context and cursor offset

* add number of pages loaded to infra.explore.load.more event instead of offset

* captures infra.explore.metric.{added, removed, aggregationChange} events

* tracks infra.explore.metric.{added, removed, aggregationChanged} for groups in infra explore"

* differentiates pre-event infra explore state props from event props

* clean up explore tracking code

* consolidate querybuilder tracking configuration into objects to reduce component api pollution

* fix pr review comments

* Review Infra Explore Mixpanel Tracking (#4540)

# Why

During the review I noticed a unusual proptype / default value pattern.

Normally I wouldn't bother as much about this, but this PR is establishing
a new pattern that will likely get adopted in more product areas in the
future.

# What

With this PR I want to propose to avoid the creation of fallback objects.
We can avoid the creation of these fallback objects and instead incur a
tiny perf penalty when calling the tracking functions via opt-chaining.

This also has the benefit that the tracking function call signature is in
line with the proptype definition. Without this change, the tracking function
calls looked as if they might run into errors due to them being optional. Only
at closer inspection did it become evident that some components within the
React tree establish defaults (with default trackers) and hence no errors occur.

Co-authored-by: Ben Blackmore <bripkens@gmail.com>