Establish forward compatibility for API token CRUD (#4546)
# Why

In a follow up release we are going to start differentiating between
API tokens' internal technical IDs and their access granting tokens.
The UI needs to be able to work with both old and new APIs / models
so that we don't have an end-user noticeable impact.

# What

Rework API token CRUD so that we provide the expected IDs to the
backend. This happens through fallbacks. Whenever possible we send
the new kind of IDs (`internalId`) to the backend. When this one is
not available, we will use the old one (`id`). This in turn means that
the UI is forward compatible with the next release.

To properly verify this as part of the release preparation, the E2E
test suite [was extended with E2E tests for API token CRUD](https://github.com/instana/backend/pull/7921).

Furthermore, this PR removes the usage of immutableJS. This only creates
confusion within this code path. In the future we want to avoid all
immutableJS usage.