Default to lazy loading widgets (#4578)
# Why

Custom dashboards are usable by our customers for some time now.
Similar to what we see in Grafana, customers do tend to create large
dashboards that require scrolling. These large dashboards put
load on the system. Some of this load is unnecessary because our
end-users are not always scrolling down to see all the content.

# What

Change the custom dashboard default behavior to lazy loading/rendering
widgets only when they are within the viewport. This will reduce
requests to those that are actually necessary.

It furthermore allows us to raise the limit for the number of widgets
in custom dashboards without performance impact.