Support custom dashboard "View as JSON" (#4588)
# Why

It is often the case that one does not have edit access to
custom dashboards. In these cases it can happen that one
would like to inspect the custom dashboard configuration,
e.g., to prepare for custom dashboard configuration via our
public API. Currently, it is not possbile to at all to
inspect the widget configuration.

# What

Let's add a simple `View as JSON` mode when the users does not
have edit access to a dashboard. This helps address some of the
use case, even if that might be fully convincing.