Use the initial state instead of the stale state when deps change (#4614)
# Why

Switching between UA2 views _Calls (v2 beta)_ and _Traces (v2 beta)_ leads to an error. The error can be reproduced on test, but not when running ui-client locally. 

The reason for this error is that the `useCursorPagination` hook returns a stale `getCalls` result instead of expected `getTraces` result.

The `useCursorPagination` hook resets the `state` to the initial value using the `useEffect` hook, when any of the `useCursorPagination` hook's dependencies change. However, since the `useEffect` hook's callback is executed after the render phase, the `useCursorPagination` hook returns a result based on the old `state`.

# What

- Storing the previous dependencies.
- The previous state will be used only if the dependencies didn't change, otherwise the initial state will be used.

Unfortunately, I didn't manage to test this using the React Hooks testing library, because resetting the state in the `useEffect` hook triggers another rerender, which means that the `useCursorPagination` hook gets called a second time, returning an expected result based on the initial state this time. Checking the `result.current` in the test after calling the `rerender` function therefore misses the temporar inconsistent state. The best way to test this is by adding debugging log messages.