Immediately return URL string instead of an observable (#4635)
# Why

The URL state HOC and now the hook always allowed creation of
state changes as URLs. At the time, this URL wasn't returned as a
`string`, but instead as a `Observable<String>`. This however is
unnecessary as a re-render is triggered whenever the URL changes.
In-turn, this causes at least two re-render cycles.

 1. Location is updated and generates a new `Observable<String>`.
 2. `Observable<String>` is bound using `useObservable`.
 3. The value from `Observable<String>` is accessed.

# What

Immediately return the URL string from the "state change as URL"-API.