Support QB2 for websites, mobile apps and APs in custom dashboards (#4625)
* Support QB2 for websites, mobile apps and APs in custom dashboards

# Why

To align with QB2 in analyze etc. :)

# What

Handle the translation from old `tagFilter[]` to the new tag filter expression
tree when editing a widget. The way this works is not entirely ideal given
the lack of support for asynchronous translations of form models within the
custom dashboard API. However, I see the need for these asynchronous
modifications as something very uncommon. I therefore chose not to extend the
API surface area of custom dashboards. Instead, it is implemented via React
hooks using the currently existing APIs.

* Review