Automatically add a conjunction when adding a tag filter
# Why

It is currently a default behavior that a newly added tag filter will
always result in validation errors due to missing conjuctions.
Considering that a very common use case is drilling down to more specific
information, we can consider automatically adding an `AND` conjuction.
This way usage of the query builder will become a bit more intuitive.

# What

Automatically add a conjunction when adding a new tag filter (when sensible).