Add beta badges (#4656)
# Why

With the decision to avoid web and mobile app QB2 inclusion in the first
step of the beta program, we need to outline much more explicitly that
QB2 is in beta. This will also help communicate the state of this part
of the product to non-Instana champions.

# What

Add a beta badge to the data source selector and a larger variant next
to the selected data source when the data source is part of the beta
program.