Show the last result if the UA2 tag filter expression is not valid (#4712)
# Why

If the UA2 tag filter expression is not valid, a loading indicator will be shown instead of the result table and a chart. This is very annoying, since the tag filter expression will be temporarily invalid during normal editing process. 

# What

Keep the last valid result and display it if the expression is invalid.

# References

- https://instana.kanbanize.com/ctrl_board/66/cards/31254/details/