Fixing no longer workign UnifiedMetricsChart in live mode (#4722)
# Why

`UnifiedMetricsChart` does no longer work in live mode. Instead of a chart a loading indicator is shown. 


This issue is caused by the fact that the object `timeConfigExtendedForLiveMode`, which is used as a hook dependency,  changes on every re-render even when its fields do not change. This leads to an infinite re-render loop.

This issue was introduced with https://github.com/instana/ui-client/pull/4690.

# What

Making sure that the same `timeConfigExtendedForLiveMode` object is reused, if non of its fields changed.

# References

- https://instana.slack.com/archives/GA8SHDH99/p1606912221116700