Too many tag catalog calls (#4723)
# Why

When UA2 is loaded, we see 7 http requests to `/application-monitoring/catalog` endpoint for fetching the tag catalog. 
![image](https://user-images.githubusercontent.com/513511/101000753-98b30880-355e-11eb-8e14-719e7663fc08.png)

In theory for a selected timeframe and a selected datasource, it'd be sufficient to emit 2 requests to `/catalog` endpoint: one for filtering `useCase=FILTERING` and one for grouping `useCase=GROUPING`. 

The results of these 2 requests should be cached in the `ui-client` and we should only emit new requests if the timeframe changes or a new datasource is selected. 

# What
There are actually 3 issues behind:

### 1. The default datasource was set to `TRACES` in `AnalyzeView.js`:
The trace catalog was first requested before switching to `dataSource=CALLS`. Call catalog is then requested. To fix the issue, we can set `calls` as the default datasource. 

### 2. The `GroupedList` component used the non cached `getApplicationTagCatalog` request.
To fix the issue, memorize and export the cached `getGroupingCatalog` from `GroupingConfigurator` and use it in `GroupedList`.

### 3. `getApplicationTagCatalog` was called with 2 slightly different time config in a short interval, the first cached request isn't reused.
This is because `Analyze.js` is wrapped in a `ApplicationAnalyzeViewWithFixatedTimeConfig` which adjust the time config from `to=null` to a fixed timestamp
To fix the issue, we can generate the `getApplicationTagCatalog` cache id using the from timestamp rounded to week. This is also done in the `getApplicationTagCatalog` request payload to optimise caching on the server side.