Component identification is incomplete (#4800)
# Why

`WithEmptyStateFallback` has a prop called `FallbackComponent` which
can accept a React element instead of React components. This is
confusing and also causing production issues right now. The component
check is incomplete, because it does not detect the various component
definition permutations that React allows for (missing are at least
classes and components supporting refs).

# What

Keep the API unmodified for now, but fix the broken identification
through React's integrated `isValidElement` API.