Rework QB2 Custom Dashboard Support (#4880)
# Why

QB2 support in custom dashboards is non-trivial to understand
because of the difference between the persistence/query and the UI
form format. This gets even harder due to the existence of old
widget configurations leveraging `tagFilter[]`.

This also currently contains bugs where chart widgets leveraging
`tagFilter[]` are considered invalid until the dataset is expanded
in the editor (and then it magically turns valid).

# What

Rework the QB2 support and share as much code as possible across the
various product areas. Furthermore introduce a dedicated widget
migration facility to support simpler implementations for QB1 -> QB2
configuration migrations.