Set tag entity when creating a filter based on the grouping tag (#4885)
# Why

When grouping by a tag with entity set to `SOURCE`, the nested calls list is incorrect. In the picture below we can see no results even though 16,137 are expected. The same issue occurs, when clicking on 'Focus on this group' link.

The root cause of this issue is that the grouping tag entity is ignored when creating a filter for the nested list.

# What

Setting the grouping tag entity when creating a tag filter based on the grouping tag.

# References

- https://instana.kanbanize.com/ctrl_board/66/cards/37922/details/