useDisabledBodyScroll doesn't support multiple usages (#4895)
# Why

`useDisabledBodyScroll` currently fails with two components using it
a the same time. In this scenario, it can happen that one component
unmounting already re-enabled body scrolling prematurely.

# What

Apply a `react-side-effect`-like strategy to `useDisabledBodyScroll`.
Avoid `react-side-effect` usage, because it uses deprecated React APIs
and instead re-implement `react-side-effect` as a hook.