Tag filter value input text selection doesn't work in Firefox (#4995)
# Why

User reported that it is not possible to select part of a tag filter value using a touchpad or a mouse, because the whole tag filter component is draggable. This has been addressed in https://github.com/instana/ui-client/pull/4946. However, this solution doesn't work in Firefox. See references for more info.

# What

Toggle the `draggable` attribute of the Tag component's main element when entering or leaving its key or value subelements. This solution seems to work in both Chrome and Firefox.

# References

- https://bugzilla.mozilla.org/show_bug.cgi?id=800050
- https://stackoverflow.com/questions/21680363/prevent-drag-event-to-interfere-with-input-elements-in-firefox-using-html5-drag
- https://instana.zendesk.com/agent/tickets/16971
- https://instana.kanbanize.com/ctrl_board/66/cards/38125/details/