Backwards Compatible URLs in UA2 (#5169)
* Backwards Compatible URLs in UA2

# Why

The URL structure for UA2 is slightly different than it is in UA1. We need
to ensure that all old and existing hyperlinks continue to work.

# What

Render a redirect component when the URL looks like a UA1 URL. Use a
`react-router` `Redirect` to update the URL.

* Updated URL transformations

* Backwards Compatible URLs in UA2

# Why

The URL structure for UA2 is slightly different than it is in UA1. We need
to ensure that all old and existing hyperlinks continue to work.

# What

Render a redirect component when the URL looks like a UA1 URL. Use a
`react-router` `Redirect` to update the URL.

* Updated transformChartedMetricsParameters logic

* Adding chartedMetrics in the url for detailed view as well

* Removed old files

Co-authored-by: Syam Pillai <syam.sp0965@gmail.com>