Support backend latency metric formatter (#5223)
# Why

In web and mobile UA2 metric formatter is taken from the metric catalog. In order to display latency metric values "0" as "< 1ms" a new `LATENCY` formatter will be introduced in https://github.com/instana/backend/pull/8907. This new backend formatter have to mapped to the right formatter in UI client.

# What

Mapping the `LATENCY` backend formatter to the `latency` UI formatter.

# References

- https://instana.kanbanize.com/ctrl_board/66/cards/39697/details/
- https://github.com/instana/backend/pull/8907