Move fetching of endpointTypes from the Summary component to its parent, to avoid temporary undefined types when switching between different tabs on the AP dashboard (#5207)
# Why

Switching between "Services" and "Summary" tab on AP dashboards freezes the UI. This happens only for APs without HTTP calls, e.g. [Demo Database](https://test-instana.pink.instana.rocks/#/application;appId=vuo8dVg_TDKxQG07hTGflQ/summary;callsTab=call;latencyTab=overTime).

The "Calls" widget on AP dashboards shows different tabs, depending on which endpoint types are present in the AP. For different endpoint types, different tabs are marked as default. These defaults are then stored as matrix parameters in the URL.  

The endpoint types are fetched in the `Summary` component using the `useObservable` hook. When switching from the "Summary" to "Services" tab, the `Summary` component gets unmounted and the fetched endpoint types get lost. When we switch from the "Services" tab back to the "Summary" tab, the `Summary` component gets mounted again and the endpoint types get refetched. Since the endpoint type fetching is asynchronous, the component first sees it as `undefined` and shortly after it is changed to the real value. Because of this, the tab selection in the URL is updated in quick succession which triggers the update loop.

# What

- Moving fetching of `endpointTypes` from the `Summary` component to `ApplicationDashboard` component. This will prevent refetching endpoint types when switching between AP dashboard tabs.

# References

- https://instana.kanbanize.com/ctrl_board/66/cards/44338/details/