Transform UA1 orderBy matrix param to UA2 orderBy and orderByGroups matrix params (#5282)
# Why

In UA1 website and mobile ordering configuration was stored in matrix parameters `orderBy` and `orderDirection`. The same matrix parameters were used for ungrouped and grouped views.

In UA2 ordering of ungrouped and grouped views can be controlled separately, for higher flexibility of nested lists inside of groups. The configuration is stored as an object with two fields `by` and `direction`, in two different matrix parameters `orderBy` and `orderByGroups`.  

# What

Set both UA2 matrix parameters `orderBy` and `orderByGroups` when transforming UA1 ordering configuration.

# References

- https://instana.kanbanize.com/ctrl_board/66/cards/45902/details/