Refactoring CountHeader to make displaying of result counts consisten… (#5405)
# Why

Result headers in UA2 are not consistent across data sources. With introduction of faceted search which shows matching row counts for suggestions, the number of results shown can become very confusing when historic data are queried. The result count shows number of retained rows, which matches the number of rows in the result table, however, the suggestion counts show approximate total numbers, which are much higher. The same issue exists for charts showing sums, e.g. total number of calls. 

# What

- The same component displaying result counts is used by all UA2 data sources, including infra explorer.
- When historic data are queried the total represented item count as well as retained item count will be shown.
- Redundant "No grouping" label will no longer be shown.
- When grouping is applied, only data source that support showing number of results in all groups will show this information below the number of groups. 

# References

- https://miro.com/app/board/o9J_kqLK2AE=/?moveToWidget=3074457354913678631&cot=14 [bug-29] [bug-30]
- https://instana.kanbanize.com/ctrl_board/66/cards/45902/subtasks/