Excluding fixed fields during conversion of UA1 'metrics' parameter (#5459)
# Why

For some data sources in UA1 the `beaconCount` metric could be explicitly selected, even though it was always implicitly added into the result table. As seen in the first screenshot, the same metric was displayed twice, even though with different labels. In UA2 the metric labels are taken from the metric catalog, therefor the two metric columns will have the same label, which makes the duplication even more visible. 

Since the `beaconCount` metric is not supported by the _Select metrics_ dropdown in UA2, it throws the following error after clicking on it:
```
TypeError: Cannot read property 'aggregations' of undefined
    at mobileApps.b9c43ca071ff2b242eb2.js:6
    at S (mobileApps.b9c43ca071ff2b242eb2.js:6)
    at Ka (index-4d086038f2.js:6)
    at Cs (index-4d086038f2.js:6)
    at bc (index-4d086038f2.js:6)
    at hc (index-4d086038f2.js:6)
    at sc (index-4d086038f2.js:6)
    at index-4d086038f2.js:6
    at t.unstable_runWithPriority (index-4d086038f2.js:6)
    at Gi (index-4d086038f2.js:6)
```

# What

During UA1 to UA2 url conversion the fixed fields are removed from the list of selected fields (metrics). 

# References

- https://instana.kanbanize.com/ctrl_board/66/cards/45902/subtasks/ [bug-11]