Fixing broken UA2 grouped charts (#5500)
# Why

All groups in grouped charts have the same values. This is a regression introduced with https://github.com/instana/ui-client/pull/5450.

# What

The root cause of the original issue described in https://github.com/instana/ui-client/pull/5450, was a missing grouping tag catalog which wasn't passed to the `addGroupingCriteriaToFormModel` function in the `onChartableDataSeriesChange` callback. 

# References

- https://instana.kanbanize.com/ctrl_board/66/cards/45902/subtasks/ [bug-46]