Use a more specific metric 'onLoadTime' instead of 'beaconDuration' (#5507)
# Why

When jumping from the onLoad time chart widgets in the Website dashboard _Speed_ tab to UA2, the chart shows the _beacon duration_ metric instead of the expected _on load time_ metric. The same applies for the selected metrics in the result list.

In UA1 the metric label in the chart and the result list showed the _on load time_ metric, even though in the back-end query the _beacon duration_ metrics was actually used.

# What

Use the `onLoadTime` metric instead of `beaconDuration`. Both metrics should actually be interchangeable, the former has a more specific name for the `pageLoad` beacon type though. 

# References

- https://instana.kanbanize.com/ctrl_board/66/cards/45902/subtasks/ [bug-36]