Show a loading indicator while the grouped chart is loading including… (#5548)
# Why

The grouped chart component in UA2 websites and mobile apps was behaving strangely after the tag filter expression, grouping or group ordering was modified. The "No data available" indicator was sometimes shown temporarily before a loading indicator appeared. Also, in some cases after the grouped chart was loaded and displayed, it was replaced with a loading indicator again and then another grouped chart with different groups was eventually displayed.

# What

- Reseting the `chartableDataSeries` state
- The Chart component will display a loading indicator while the parent component is loading the chart configuration, i.e. `chartableDataSeries`.
- The Chart component won't start executing back-end queries, until the final chart configuration is ready.

# References

- https://instana.kanbanize.com/ctrl_board/66/cards/45902/subtasks/ [bug-37]