Resolve workaround of falling back to filters object to scope the application/service when no search term was used (#5720)
This workaround was introduced in #5589 and was only needed because the backend did not support it at that time.
It had the downside that in the S/E-selection of AP SmartAlerts, we could e.g. not search for synthetic endpoints.
This is now being resolved in the backend. Consequently, we can resolve this workaround.