Bugfix 48101 smart alerts dialog show error steps (#5867)
* fix(Smart Alert Dialog): show which steps have errors

Adds the ability to indicate which steps have configuration errors in
them in advanced mode.
This allows the user to easily see where the configuration needs to be
adjusted

Also slightly increases the size of the indicator icons to improve
readability

ISSUE: 48101

* fix(smart alert form): improve threshold form state dislapy

The Trigger step of the smart alert form did not correctly display its
validation state and showed its initial state as invalid.
Additionally threshold updates of the ap smart alert form did not correctly preserve
the touched state when being updated. This resulted in form validation
not behaving correctly.

Added handling to display the initial state of the trigger step as not
filled, instead of invalid.
Added handling to preserve the touched form state of staticThreshold
inputs on updates in the case of ap smart alerts

ISSUE: 48101

* style(SmartAlertDialog): remove unneccessary color props

Color does not change anything for svg icons.