[containers] New way of building ui-client component container images (#5911)
# Why

This is part of the same journey of moving away from building our ui-client and backend container images in [artifacts](https://github.com/instana/artifacts), that has been started in the backend as illustrated by the following backend PRs in order:

https://github.com/instana/backend/pull/9189
https://github.com/instana/backend/pull/9538
https://github.com/instana/backend/pull/9630
https://github.com/instana/backend/pull/9654
https://github.com/instana/backend/pull/9675

# What

We're replicating what's been done in [artifacts](https://github.com/instana/artifacts/blob/master/Rakefile#L542-L586) supporting both `ui-client` (source map files removed) and `ui-client-saas` container images.

The bash scripts in `containers/scripts/*` are mostly the same as the ones in the `backend`, but with some modifications to support specific `ui-client` needs. When the `ui-client` is _finally_ combined with the `backend` as a real monorepo (if that is still happening), we will obviously be combining both sets of scripts.

This is not _yet_ integrated into the `ui-client` pipeline. That will be the next step.

# References

- [Story](https://instana.kanbanize.com/ctrl_board/75/cards/46405/details/)