Fix visibility checks in TV mode (#6056)
# Why

When TV mode is active, we are leveraging a custom scroll zone in a
`position: fixed` DOM node spanning the whole viewport. Some TV mode
content, e.g. custom dashboards, leverage `react-on-screen` to identify
when lazy-load elements. This lazy loading behavior is currently broken,
because `react-on-screen` only supports scroll detection for the whole
screen.

# What

Remove `react-on-screen` and replace it with a new library that supports
scroll identification for custom elements. This also has the benefit
of replacing a more or unless unmaintained library (`react-on-screen`).
Furthermore, we can leverage a more modern and efficient browser API
called [IntersectionObserver].

[IntersectionObserver]: https://caniuse.com/intersectionobserver