Fixing UA2 charts for calls & traces which may ignore sorting and show incorrect data (#6098)
# Why

UA2 charts for calls and traces may show wrong data In some cases. This happens when the charted metric is not present in the result table, in which case the charting component calls `getUnifiedMetrics` to get the needed metric data without applying the correct sorting configuration. Depending on the sorting configuration the chart can display data for some random 5 groups incorrectly label with the top 5 group names from the result table.

If the charted metric is present in the result table, the charting component uses metric data fetched by the `getCallGroups` or `getTraceGroups` commands, which correctly applies the sorting configuration.

Unfortunately, passing the right sorting configuration to the `getUnifiedMetrics` endpoint would not resolve the issue completely, because this endpoint does not support sorting by "group name" and "earliest timestamp".

# What

Using the same mechanism UA2 charting common component use, i.e., metric data will be fetch for each of the 5 groups separately, by extending the tag filter expression with a filter on the given group name.

Please note that this is just a temporary fix, since the UA2 calls & traces are going to use the UA2 common components, see https://instana.kanbanize.com/ctrl_board/66/cards/48287/details/. 

# References

- https://instana.kanbanize.com/ctrl_board/66/cards/55595/details/