Move storybook dependencies to a sub-directory (#6255)
# Why

While trying to update our dependencies, I once again ran into
issues with Storybook. Storybook brings a whole slew of dependencies
that are causing resolution conflicts with our own/desired dependencies.

For example, we need to upgrade react-select. The latest version is ^4.
We are currently on ^1. Storybook wants ^3. If we would upgrade, we would
want to upgrade to the latest version (instead of upgrading to an outdated
version).

This is just one out of many dependencies that are causing problems, because
they are existing within Storybook and ui-client. In theory, we could upgrade
the desired dependency and Storybook within one step, but this would result
in an extremely huge and complicated change.

Going forward, we need to find a way to decouple Storybook from ui-client to a
certain degree.

# What

Move the Storybook dependencies into a separate `package.json` located within
the `storybook` directory. This way, the dependencies are somewhat better
decoupled. It leaves something to be desired, because the dependency conflicts
continue to exist.

However, the individual upgrade/migration steps become smaller and we can
upgrade dependencies step by step.