fix(SmartAlertDialog): fix unnecessary baseline resets (#6343)
useSmartAlertFormSideEffects would also trigger changes when form fields
were added or removed. This caused unnecessary baseline resets which
resulted in the same baseline. Thus not updating the form with the new
baseline again.
Now this behaviour is configurable in useFormSideEffects and
useSmartAlertFormSideEffects only triggers on field changes not
additions or deletions.