Ensure 3.xxx.xxx images have the 'com.instana.image.tag' LABEL that matches the actual image tag (#10562) (#6396)
# Why

While doing backwards compatibility testing with the new 3.xxx.xxx image, I discovered that the com.instana.image.tag LABEL in the image isn't properly set when we retag the ui-client images from the backend pipeline (and vice versa, when we retag the backend images from the ui-client pipeline). We retag images to ensure that there's a single image version that's applicable across all backend and ui-client images since instanactl only understands a single version number.

For example, in the ui-client pipeline:

When the ui-client pipeline runs and creates image version `3.202.376-0`, that contains `backend: 1.202.771, ui-client: 1.202.376`, this means it takes backend image version `3.202.771-0` and retags it as `3.202.376-0`. However, the image labels themselves have already been created (and are static), so the backend images will still have `LABEL com.instana.image.tag = 3.202.771-0` even though the actual image tag is `3.202.376-0`

So instead of just "retagging" the backend images in the ui-client pipeline (and vice versa), we'll need to build a new version of the backend images from the existing version and override that label with the new version.

# What

Build a new version of the backend images from the current version and add another layer that overrides the `com.instana.image.tag LABEL` value to set to the new version

# References

- [Story / Card](https://instana.kanbanize.com/ctrl_board/75/cards/60407/details/)