Update proxrox for Cloudflare compatibility (#6398)
# Why

The dev mode of ui-client against release-instana is resulting in 404s.

We are in the process of having Cloudflare in front of our environment.
Cloudflare needs additional information to correctly route traffic that
our dev mode proxy is not setting at this point. As a result, Cloudflare
fails to correctly identify the customer/upstream and responds with a 404.
This in turn breaks the UI dev mode.

# What

Upgrade proxrox to the latest version. In the [latest version], proxrox
enables TLS SNI support. This in turn fixes Cloudflare routing and through
this the development mode.

For more information how Cloudflare comes into play with ui-client and our
product, please see the prepared [UI weekly sync notes].

[latest version]: https://github.com/bripkens/proxrox/commit/e914de11e89900ea447ad6c299891701c6636ef7
[UI weekly sync notes]: https://www.notion.so/instana/2021-06-01-90d91d8d4f454f0a876ad22eb982eb30