Ensure backend + ui-client images use a single version source for 3.xxx.xxx instead of co-opting its existing tar.gz version (#6422)
# Why

Currently each of the ui-client and backend pipelines co-opt the version of the tar.gz (1.xxx.xxx) to construct the version of the image by replacing the first number 1 -> 3, and adding a -0 at the end. For example: 1.203.89 will be converted to 3.203.89-0.

Though the versions of the images created by the ui-client pipeline never override the versions of the images created by the backend pipeline, it is confusing for the numbers to flip-flop depending on which pipeline created them.

For example:
The ui-client pipeline created all images with version 3.203.379-0 but when the backend pipeline ran, it created all images with version 3.203.777-0. The next run of the ui-client pipeline run will create all images with version 3.203.380-0 which is technically more recent than 3.203.777-0 because it has a more recent version of the ui-client code but because 380-0 is smaller than 777-0, it is confusing to tell from the version which one is newer.

# What

- Use a single version source for the 3.xxx.xxx images via `ci-shared-tools/componentVersioning/getInstanaImageVersion.js`
- Lookup latest stable Instana image version built by the backend pipeline via the 'stable' `instana-image-from-backend` version in the EFS versioning system in order to determine which version of backend images to rebuild
- Mark that version as a 'stable' `instana-image-from-ui-client` version in the EFS versioning system (so that it can be looked up by the respective backend branch pipeline)
- Remove all references of `iteration` in the container definition files and just reference an `image_version` that will be in the format `3.xxx.xxx-0`

# References

- [Story / Card](https://instana.kanbanize.com/ctrl_board/75/cards/60522/details/)