Build additional OpenShift-compatible images (#6481)
Similar to https://github.com/instana/backend/pull/10736 + https://github.com/instana/backend/pull/10776/commits/63e97ecd7ab65bc0941a94aaf4039673f57c8e61

# Why

In order to get Instana running on IBMCloud/ROKS, we have to make additional changes to the image definitions for butler, ingress, ingress-global. However, these changes are not going to work on some of Instana's existing OnPrem installations.

# What

Modify current scripts to build an additional image for ui-client with an `-openshift` tag suffix. In the backend repo, the scripts will use the `container.openshift` file for butler, ingress and ingress-global but will fall back to the previous `container` file that was defined previously (either a custom one provided by the component, or the default one), but since in this repo the ui-client does not have a `container.openshift` file, it'll just use the existing `container` file.

Also make sure to retag the backend images with the `-openshift` version generated from this repository. Similar to the follow-up commit that was made in the backend :point_up: