Enforce intended package architecture (#6533)
# Why

One of the core topics codeowners should review as part of
ui-client PRs are cross-package imports. For example, we should
ensure that in-components doesn't have in-application or
in-forge imports. In the short run, this protects us from wrong
technical designs. In the long run, this ensures that we have a
way to split up ui-client into independently managed and
releasable packages ("Microfrontend architecture").

If we don't pay attention to this now, then we are going to have
problems down the road when we want to introduce an
architecture change.

# What

Establish a mechanism for per-package configuration of the
`no-restricted-imports` rule. Give it a try with the
`in-components` package. This change currently fails linting.
It is meant as a starting point for discussions and follow-up
refactorings.