Use Intl based Date formatters (#6663)
# Why

We recently moved all  data formatting capabilities over to
`@instana/format-date`. Not only was this a code move, but
we also added support for out-of-the-box i18n through usage
of the standard JavaScript Intl APIs, alternative date/time
presentation formats and support for all time zones.

# What

Use the new APIs.

We do not yet leverage the alternative date/time zone formatting
capabilities (to reduce impact/risk).