in-server authentication code has uncaught promise rejections (#6899)
# Why

Within the in-server code, we face uncaught promise rejections whenever users have a valid seeming cookie pointing to an expired session. In those cases, the subRequestPromises are never handled, because we return in the function before the promises are listened to.

# What

Make sure that we always add a rejection handler.