Remove findDOMNode (#7401)
# Why

`findDOMNode` is deprecated.

# What

There was a lot of attempts already to remove `findDOMNode` from the chart
implementation. Unfortunately, they all had bugs that influenced the
chart resize behavior. Looking into these changes, it can be observed that
all approaches modified the DOM and/or CSS structure.

These changes are actually not necessary. The only necessary change is that
we properly forward React refs to re-implement the behavior of `findDOMNode`.