Read only AP configs (#7361)
# Why

Users that are not allowed to configure APs should at least be able to see the configuration of APs that they have access for. Customers gave feedback that their users have difficulties to put the data they are viewing in proper context.

# What

Create a read-only AP configuration view for users that do not have the `canConfigureApplications ` permission. Before, the configuration tab was not visible at all. The readonly view consists of:

- No more input field for the AP name but rather a simpe label
- Read only QueryBuilder for filter of AP
- No more checkboxes for scope and boundaryScope, but rather a textual representation of the selected option
- Remove delete button alltogether in read-only mode
- Info box at the top to indicate that the user is viewing the config in read-only mode

This PR is also rewrites `Removal.js` from a class component to a functional component.